refactor(netutil): consolidate gateway discovery into single route.go
- Merge platform-specific route files into unified route.go
- Use gateway package for cross-platform default gateway discovery
- Simplify code by removing build tag conditional logic
- Preserve environment in sudo command (Taskfile.yaml)
